+++ librustzcash/zcash_client_sqlite/src/wallet/commitment_tree.rs
@@ -1661,8 +1661,6 @@ mod tests {
             Network::TestNetwork,
             test_clock(),
             test_rng(),
-            #[cfg(feature = "zakura-pir-enhance")]
-            zcash_client_backend::data_api::enhance_pir::EnhancementMode::Standard,
         )
         .unwrap();
         data_file.keep().unwrap();
@@ -1782,8 +1780,6 @@ mod tests {
             Network::TestNetwork,
             test_clock(),
             test_rng(),
-            #[cfg(feature = "zakura-pir-enhance")]
-            zcash_client_backend::data_api::enhance_pir::EnhancementMode::Standard,
         )
         .unwrap();
         WalletMigrator::new().init_or_migrate(&mut db).unwrap();
@@ -1898,8 +1894,6 @@ mod tests {
             Network::TestNetwork,
             test_clock(),
             test_rng(),
-            #[cfg(feature = "zakura-pir-enhance")]
-            zcash_client_backend::data_api::enhance_pir::EnhancementMode::Standard,
         )
         .unwrap();
         data_file.keep().unwrap();
@@ -2013,8 +2007,6 @@ mod tests {
             Network::TestNetwork,
             test_clock(),
             test_rng(),
-            #[cfg(feature = "zakura-pir-enhance")]
-            zcash_client_backend::data_api::enhance_pir::EnhancementMode::Standard,
         )
         .unwrap();
         data_file.keep().unwrap();
@@ -2134,8 +2126,6 @@ mod tests {
             Network::TestNetwork,
             test_clock(),
             test_rng(),
-            #[cfg(feature = "zakura-pir-enhance")]
-            zcash_client_backend::data_api::enhance_pir::EnhancementMode::Standard,
         )
         .unwrap();
         data_file.keep().unwrap();