+++ librustzcash/zcash_client_sqlite/src/wallet/init/migrations.rs
@@ -845,8 +845,6 @@ pub(crate) mod tests {
             Network::TestNetwork,
             test_clock(),
             test_rng(),
-            #[cfg(feature = "zakura-pir-enhance")]
-            zcash_client_backend::data_api::enhance_pir::EnhancementMode::Standard,
         )
         .unwrap();
 
@@ -868,8 +866,6 @@ pub(crate) mod tests {
             Network::TestNetwork,
             test_clock(),
             test_rng(),
-            #[cfg(feature = "zakura-pir-enhance")]
-            zcash_client_backend::data_api::enhance_pir::EnhancementMode::Standard,
         )
         .unwrap();